Src/Main_Algorithm/requirements.txt
Django
kafka-python==2.0.2
singleton-decorator==1.0.0
mysql-connector-python==8.3.0